+++ packages/isaacscript-common/src/objects/roomShapeToDoorSlotsToGridIndexDelta.ts
@@ -1,5 +1,7 @@
 import { DoorSlot, RoomShape } from "isaac-typescript-definitions";
 import { LEVEL_GRID_ROW_WIDTH } from "../core/constants";
+import { newReadonlyMap } from "../functions/map";
+import { HasAllEnumKeys } from "../types/HasAllEnumKeys";
 
 const LEFT = -1;
 const UP = -LEVEL_GRID_ROW_WIDTH;
@@ -10,11 +12,9 @@ const DOWN = LEVEL_GRID_ROW_WIDTH;
  * Deltas are considered to be from the safe grid index of the room (i.e. the top left corner, or
  * top right corner in the case of `RoomShape.LTL`).
  */
-export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
-  readonly [key in RoomShape]: Map<DoorSlot, int>;
-} = {
+export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA = {
   // 1
-  [RoomShape.SHAPE_1x1]: new Map([
+  [RoomShape.SHAPE_1x1]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT], // 2
@@ -22,19 +22,19 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 2
-  [RoomShape.IH]: new Map([
+  [RoomShape.IH]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.RIGHT_0, RIGHT], // 2
   ]),
 
   // 3
-  [RoomShape.IV]: new Map([
+  [RoomShape.IV]: newReadonlyMap([
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.DOWN_0, DOWN], // 3
   ]),
 
   // 4
-  [RoomShape.SHAPE_1x2]: new Map([
+  [RoomShape.SHAPE_1x2]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT], // 2
@@ -44,13 +44,13 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 5
-  [RoomShape.IIV]: new Map([
+  [RoomShape.IIV]: newReadonlyMap([
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.DOWN_0, DOWN + DOWN], // 3
   ]),
 
   // 6
-  [RoomShape.SHAPE_2x1]: new Map([
+  [RoomShape.SHAPE_2x1]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT + RIGHT], // 2
@@ -60,13 +60,13 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 7
-  [RoomShape.IIH]: new Map([
+  [RoomShape.IIH]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.RIGHT_0, RIGHT + RIGHT], // 2
   ]),
 
   // 8
-  [RoomShape.SHAPE_2x2]: new Map([
+  [RoomShape.SHAPE_2x2]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT + RIGHT], // 2
@@ -78,7 +78,7 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 9
-  [RoomShape.LTL]: new Map([
+  [RoomShape.LTL]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, DOWN + LEFT + UP], // 1
     [DoorSlot.RIGHT_0, RIGHT], // 2
@@ -90,7 +90,7 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 10
-  [RoomShape.LTR]: new Map([
+  [RoomShape.LTR]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT], // 2
@@ -102,7 +102,7 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 11
-  [RoomShape.LBL]: new Map([
+  [RoomShape.LBL]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT + RIGHT], // 2
@@ -114,7 +114,7 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
   ]),
 
   // 12
-  [RoomShape.LBR]: new Map([
+  [RoomShape.LBR]: newReadonlyMap([
     [DoorSlot.LEFT_0, LEFT], // 0
     [DoorSlot.UP_0, UP], // 1
     [DoorSlot.RIGHT_0, RIGHT + RIGHT], // 2
@@ -124,4 +124,4 @@ export const ROOM_SHAPE_TO_DOOR_SLOTS_TO_GRID_INDEX_DELTA: {
     [DoorSlot.RIGHT_1, DOWN + RIGHT], // 6
     [DoorSlot.DOWN_1, RIGHT + DOWN], // 7
   ]),
-} as const;
+} as const satisfies HasAllEnumKeys<RoomShape>;