+++ packages/isaacscript-common/src/objects/versusScreenBackgroundColors.ts
@@ -1,12 +1,11 @@
 import { StageID } from "isaac-typescript-definitions";
+import { HasAllEnumKeys } from "../types/HasAllEnumKeys";
 
 /** We arbitrarily specify a default color equal to that of Basement. */
 const DEFAULT_COLOR = Color(26 / 255, 14 / 255, 12 / 255);
 
 /** These values are taken from StageAPI. */
-export const VERSUS_SCREEN_BACKGROUND_COLORS: {
-  readonly [key in StageID]: Color;
-} = {
+export const VERSUS_SCREEN_BACKGROUND_COLORS = {
   [StageID.SPECIAL_ROOMS]: DEFAULT_COLOR, // 0
   [StageID.BASEMENT]: Color(26 / 255, 14 / 255, 12 / 255), // 1
   [StageID.CELLAR]: Color(26 / 255, 17 / 255, 13 / 255), // 2
@@ -36,4 +35,4 @@ export const VERSUS_SCREEN_BACKGROUND_COLORS: {
   [StageID.MORTIS]: Color(13 / 255, 14 / 255, 12 / 255), // 34
   [StageID.HOME]: DEFAULT_COLOR, // 35
   [StageID.BACKWARDS]: DEFAULT_COLOR, // 36
-} as const;
+} as const satisfies HasAllEnumKeys<StageID>;